ensure that tests pass on osx
