+++ networker/utils.py
@@ -37,6 +37,19 @@ def geograph_to_json(g):
     return js_g
 
 
+def get_rounded_edge_sets(geograph, round_precision=8):
+    """
+    Get set of edges with coordinates rounded to a certain precision
+    """
+    tup_map = {i: 
+               tuple(map(lambda c: round(c, round_precision), coords))
+               for i, coords in geograph.coords.items()}
+
+    edge_sets = map(lambda e: frozenset([tup_map[e[0]], tup_map[e[1]]]),
+                    geograph.edges())
+    return set(edge_sets)
+
+
 """
 # spherical drawing helpers
 # plot maps
